Beginning support for Verilator backend. - Re-organize packages   - create backends   - add a treadle backend - Generalize backends in anticipation of verilator   - create BackendExecutive     - combinational path building should be shared so moved to here.   - new TreadleBackend     - Uses BackendExecutive     - Eliminate try block in object       - Simplifies debugging   - Split out TreadleExecutive from TreadleBackend   - ThreadedBackend     - provides dut access       - this required parameterization of trait with dut's type
